Move move logic to parent
